refactor: 테스트 코드 의존성 및 명칭 변경
--- 본문 ---
- `io.github.metdaisy.amaazon.auth.application.event.FormSignUpTask`를 `io.github.metdaisy.amaazon.user.application.event.FormSignUpTask`로 변경하여 사용자 도메인으로 이벤트 위치를 이동했습니다.
- `UserControllerTest`에서 `UserMapper` 의존성을 제거하고, `UserService`가 직접 DTO를 반환하도록 변경했습니다.
- `UserControllerTest`에서 `userService.update`를 `userService.updateProfile`로, `userService.find`를 `userService.findProfile`로 메서드 이름을 변경했습니다.
- `SocialLoginSuccessHandlerTest`에서 `AuthException` 대신 `AuthenticationCredentialsNotFoundException`을 사용하도록 변경했습니다.
- `ModularityTest`에서 `product` 도메인을 `catalog`로 변경했습니다.
- 테스트 코드의 `@DisplayName`을 한국어로 통일하고 구체적인 동작을 설명하도록 수정했습니다.
